refactor: give comments its own view models
The comments package already owns its services, but the view models
that drive the comment table still sat in the shared viewmodels and
dialogs packages. Spreading one area over three places makes it hard
to see what belongs to comments, so they now sit next to the services
they read.
