+++ mpvqc/dialogs/__init__.py
@@ -3,6 +3,5 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from .about import MpvqcAboutDialogViewModel as MpvqcAboutDialogViewModel
-from .comment_types import MpvqcCommentTypesDialogViewModel as MpvqcCommentTypesDialogViewModel
 from .edit_input import MpvqcEditInputDialogViewModel as MpvqcEditInputDialogViewModel
 from .edit_mpv import MpvqcEditMpvDialogViewModel as MpvqcEditMpvDialogViewModel
