--- mpvqc/viewmodels/views/table/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-FileCopyrightText: mpvQC developers
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-from .comment_table import MpvqcCommentTableViewModel as MpvqcCommentTableViewModel
-from .placeholder import MpvqcPlaceholderViewModel as MpvqcPlaceholderViewModel
-from .search_box import MpvqcSearchBoxViewModel as MpvqcSearchBoxViewModel
-from .table_utility import MpvqcTableUtilityViewModel as MpvqcTableUtilityViewModel
