+++ pyproject.toml
@@ -157,10 +157,16 @@ files = [
     "mpvqc/comments/services/type_validator.py",
     "mpvqc/comments/services/types_policy.py",
     "mpvqc/comments/services/view_action.py",
+    "mpvqc/comments/viewmodels/__init__.py",
+    "mpvqc/comments/viewmodels/comment_table.py",
+    "mpvqc/comments/viewmodels/comment_types_dialog.py",
+    "mpvqc/comments/viewmodels/label_width_calculator.py",
+    "mpvqc/comments/viewmodels/placeholder.py",
+    "mpvqc/comments/viewmodels/search_box.py",
+    "mpvqc/comments/viewmodels/table_utility.py",
     "mpvqc/comments/wiring.py",
     "mpvqc/dialogs/__init__.py",
     "mpvqc/dialogs/about.py",
-    "mpvqc/dialogs/comment_types.py",
     "mpvqc/dialogs/edit_input.py",
     "mpvqc/dialogs/edit_mpv.py",
     "mpvqc/enums/__init__.py",
@@ -287,7 +293,6 @@ files = [
     "mpvqc/viewmodels/message_boxes/version_check.py",
     "mpvqc/viewmodels/utility/__init__.py",
     "mpvqc/viewmodels/utility/fonts.py",
-    "mpvqc/viewmodels/utility/label_width_calculator.py",
     "mpvqc/viewmodels/utility/window.py",
     "mpvqc/viewmodels/views/__init__.py",
     "mpvqc/viewmodels/views/footer.py",
@@ -299,11 +304,6 @@ files = [
     "mpvqc/viewmodels/views/player/__init__.py",
     "mpvqc/viewmodels/views/player/new_comment_menu.py",
     "mpvqc/viewmodels/views/player/player.py",
-    "mpvqc/viewmodels/views/table/__init__.py",
-    "mpvqc/viewmodels/views/table/comment_table.py",
-    "mpvqc/viewmodels/views/table/placeholder.py",
-    "mpvqc/viewmodels/views/table/search_box.py",
-    "mpvqc/viewmodels/views/table/table_utility.py",
     "mpvqc/views/__init__.py",
     "mpvqc/views/player_framebuffer_object.py",
     "mpvqc/views/player_win_id.py",
