+++ test/comments/viewmodels/test_comment_types_dialog.py
@@ -12,7 +12,7 @@ from mpvqc.comments.services import (
     CommentTypeValidatorService,
     default_comment_types,
 )
-from mpvqc.dialogs import MpvqcCommentTypesDialogViewModel
+from mpvqc.comments.viewmodels import MpvqcCommentTypesDialogViewModel
 
 
 @pytest.fixture
