+++ test/comments/viewmodels/test_search_box.py
@@ -8,8 +8,8 @@ import inject
 import pytest
 
 from mpvqc.comments.services import CommentsService
+from mpvqc.comments.viewmodels import MpvqcSearchBoxViewModel
 from mpvqc.shared import Comment
-from mpvqc.viewmodels import MpvqcSearchBoxViewModel
 
 DEFAULT_COMMENTS_SEARCH = (
     Comment(time=0, comment_type="commentType", comment="Word 1"),
