+++ test/comments/viewmodels/test_table_utility.py
@@ -6,9 +6,9 @@ import inject
 import pytest
 
 from mpvqc.comments.services import CommentsService, TimeFormatPolicyService
+from mpvqc.comments.viewmodels import MpvqcTableUtilityViewModel
 from mpvqc.services import PlayerService
 from mpvqc.shared import Comment
-from mpvqc.viewmodels import MpvqcTableUtilityViewModel
 
 ONE_HOUR_MS = 3_600_000
